refactor(tactics): extract useTacticsLibrary hook from TacticsTab
Move the custom-tactics library flow out of TacticsTab into a dedicated
hook: its state (customTactics, activeTacticId, draftTacticName,
presetAnchorId), the load/save/preset-sync/name-sync effects (dep arrays
and the first-save scope-guard ref preserved verbatim), the derived
dirty-state and tactic library, and the create/duplicate/save/apply
handlers. The formation/play-style persistence handlers stay in the
parent and are injected, preserving applyTacticSelection's
Promise<boolean> short-circuit. Pure move.

TacticsTab.tsx: 963 -> 750 lines.
